Revert "TEST: Try switching to `Number=f32`"
This reverts commit 5577f26e15ff0c15ddd0b83108bea8e662fc285f.

+++ rayna_shared/src/def/types.rs
@@ -4,7 +4,7 @@ pub type Channel = f32;
 pub type Pixel = Rgb<Channel>;
 pub type ImgBuf = ImageBuffer<Pixel, Vec<Channel>>;
 
-pub type Number = f32;
+pub type Number = f64;
 pub type Angle = glamour::Angle<Number>;
 pub type Vector2 = glamour::Vector2<Number>;
 pub type Vector3 = glamour::Vector3<Number>;
